% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{collapsed_predict_label}
\alias{collapsed_predict_label}
\title{Predict labels given omega and Cnv}
\usage{
collapsed_predict_label(aug_data = AugmentedData(), Cnv)
}
\arguments{
\item{aug_data}{An augmented data for polya-gamma sampling.}

\item{Cnv}{A numeric matrix with dim(m_new,m), cross covariance matrix
between new sample and training sample.}
}
\value{
\code{Y_new} A numeric matrix with dim(m_new, J), each row indicates
the predictive "label" of the corresponding new sample.
}
\description{
Predict labels given omega and Cnv
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- matrix(sample.int(3*3 ,replace=TRUE),3,3)
aug_data <- AugmentedData(C,Y)
Cnv <- matrix(rnorm(5*3), 5, 3)
collapsed_predict_label(aug_data, Cnv)
}
