% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_similarity.R
\name{cross_similarity}
\alias{cross_similarity}
\title{Compute cross similarity matrix Z between X and U}
\usage{
cross_similarity(X, U, r = 3L, kernel = "lae", gl = "rw", cl = NULL)
}
\arguments{
\item{X}{A numeric matrix with dim (n,d), original sample,
each row indicates one original point in R^d.}

\item{U}{A numeric matrix with dim (s,d) or (s,d+1), sub-sample,
each row indicates one representative point in R^d,
where the d+1 column indicates the number of points in each cluster if it exists.}

\item{r}{An integer, the number of the nearest neighbor points.}

\item{kernel}{A character vectors in c("lae"), the type of kernel in computing
the cross similarity matrix Z, the defaulting value is "lae".}

\item{gl}{A character vector in c("rw", "normalized", "cluster-normalized"),
indicates how to construct the stochastic transition matrix. "rw" means random walk,
"normalized" means normalized random walk, "cluster-normalized" means
normalized random walk with cluster membership re-balance. The defaulting gl
is "rw".}

\item{cl}{he cluster to make parallel computing,
typically generated by parallel::makeCluster(num_workers).
The defaulting value of cl is NULL, that is, sequential computing.}
}
\value{
\code{Z} A numeric matrix with dim (n,s),
the stochastic transition matrix from X to U.
}
\description{
Compute cross similarity matrix Z between X and U
}
\examples{
X <- matrix(rnorm(5*2), 5, 2)
U <- subsample(X, 2, "random")
cross_similarity(X, U, 2)
}
