% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cross_similarity_lae_cpp}
\alias{cross_similarity_lae_cpp}
\title{Compute cross similarity matrix Z between X and U by local anchor embedding}
\usage{
cross_similarity_lae_cpp(X, U, r = 3L, gl = "rw")
}
\arguments{
\item{X}{A numeric matrix with dim (n,d), original sample,
each row indicates one original point in R^d.}

\item{U}{A numeric matrix with dim (s,d) or (s,d+1), sub-sample,
each row indicates one representative point in R^d,
where the d+1 column indicates the number of points in each cluster if it exists.}

\item{r}{An integer, the number of the nearest neighbor points.}

\item{gl}{A character vector in c("rw", "normalized", "cluster-normalized"),
indicates how to construct the stochastic transition matrix. "rw" means random walk,
"normalized" means normalized random walk, "cluster-normalized" means
normalized random walk with cluster membership re-balance. The defaulting gl
is "rw".}
}
\value{
\code{Z} A numeric sparse dgr matrix with dim (n,s),
the stochastic transition matrix from X to U.
}
\description{
Compute cross similarity matrix Z between X and U by local anchor embedding
}
