% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{error_rate}
\alias{error_rate}
\title{Compute classification error rates}
\usage{
error_rate(Y_new, Y_pred)
}
\arguments{
\item{Y_new}{Labels of new samples.}

\item{Y_pred}{Predicted labels of new samples.}
}
\value{
A double, the error rate.
}
\description{
Compute classification error rates
}
\examples{
Y_new <- c(1,1,0,0)
Y_pred <- c(1,0,1,0)
error_rate(Y_new, Y_pred)
}
