% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Euclidean distance between one point and reference points}
\usage{
euclidean_distance(x, U)
}
\arguments{
\item{x}{One original point, a vector of length d.}

\item{U}{Reference points, a (s,d) matrix, each row indicates one reference point.}
}
\value{
The Euclidean distance between x and U, a vector of length s.
}
\description{
Euclidean distance between one point and reference points
}
\examples{
x <- rnorm(3)
U <- matrix(rnorm(30), nrow=10, ncol=3)
euclidean_distance(x, U)
}
