% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_breaking.R
\name{f_to_pi}
\alias{f_to_pi}
\title{Stick breaking transform from f to pi}
\usage{
f_to_pi(f)
}
\arguments{
\item{f}{A numeric matrix with dim(M,J-1), latent function.}
}
\value{
\code{pi} A numeric matrix with dim(M, J), class probability.
}
\description{
Stick breaking transform from f to pi
}
\examples{
f <- matrix(rnorm(3*3),3,3)
f_to_pi(f)
}
