% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit_gl_logit_gp_rcpp}
\alias{fit_gl_logit_gp_rcpp}
\title{Fit logistic regression with GLGP}
\usage{
fit_gl_logit_gp_rcpp(
  X,
  Y,
  X_new,
  K,
  N = NULL,
  sigma = 0.001,
  a2s = NULL,
  threshold = 0.01,
  sparse = TRUE,
  approach = "posterior",
  models = list(subsample = "kmeans", kernel = "lae", gl = "cluster-normalized", root =
    TRUE),
  output_cov = FALSE
)
}
\arguments{
\item{X}{Training sample, a (m, d) matrix, each row indicates one point in R^d.}

\item{Y}{A numeric vector with length(m), count of the positive class.}

\item{X_new}{Testing sample, a (n-m, d) matrix, each row indicates one point in R^d.}

\item{K}{An integer, the number of used eigenpairs to construct heat kernel,
the defaulting value is \code{NULL}, that is, \code{K=min(n,s)}.}

\item{N}{A numeric vector with length(m), total count.}

\item{sigma}{A non-negative number, the weight coefficient of ridge penalty on H,
the defaulting value is 1e-3.}

\item{a2s}{A numeric vector, the searching range for bandwidth.}

\item{threshold}{A double, the threshold ratio for sparse GLGP, defaulting value is 0.01.}

\item{sparse}{bool, sparse GLGP or not, defaulting value is \code{TRUE}.}

\item{approach}{A character vector, taking value in c("posterior", "marginal"),
decides which objective function to be optimized, defaulting value is \code{posterior}.}

\item{models}{A list with four components
\describe{
\item{subsample}{the method of subsampling, the defaulting value is \code{kmeans}.}
\item{kernel}{the type of kernel to compute cross similarity matrix W, the
defaulting value is \code{lae}.}
\item{gl}{the kind of graph Laplacian L, the defaulting value is \code{cluster-normalized}.}
\item{root}{whether to square root eigenvalues of the two steps similarity matrix W,
the defaulting value is \code{TRUE}.}
}}

\item{output_cov}{Bool, whether to output covariance, defaulting value is \code{FALSE}.}
}
\value{
\code{res} A list with two components including \code{pars} and \code{Y_pred},
where \code{Y_pred} is also a list with two components, that is,
\describe{
\item{train}{A numeric vector with length(m), each element indicates
the label in the train data point.}
\item{test}{A numeric vector with length(m_new), each element indicates
the label in the test data point.}
}
}
\description{
Fit logistic regression with GLGP
}
