% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit_gl_logit_mult_gp_rcpp}
\alias{fit_gl_logit_mult_gp_rcpp}
\title{Fit Gaussian process logistic multinomial regression with GLGP}
\usage{
fit_gl_logit_mult_gp_rcpp(
  X,
  Y,
  X_new,
  K,
  sigma = 0.001,
  a2s = NULL,
  threshold = 0.01,
  sparse = TRUE,
  approach = "posterior"
)
}
\arguments{
\item{X}{Training sample, a (m, d) matrix, each row indicates one point in R^d.}

\item{Y}{A numeric vector with length(m), indicating the labels of multi-classes,
\code{Y} should be continuous integers, such as 0,1,2,...,9.}

\item{X_new}{Testing sample, a (n-m, d) matrix, each row indicates one point in R^d.}

\item{K}{An integer, the number of used eigenpairs to construct heat kernel,
the defaulting value is \code{NULL}, that is, \code{K=min(n,s)}.}

\item{sigma}{A non-negative number, the weight coefficient of ridge penalty on H,
the defaulting value is 1e-3.}

\item{a2s}{A numeric vector, the searching range for bandwidth.}

\item{threshold}{A double, the threshold ratio for sparse GLGP, defaulting value is 0.01.}

\item{sparse}{bool, sparse GLGP or not, defaulting value is \code{TRUE}.}

\item{approach}{A character vector, taking value in c("posterior", "marginal"),
decides which objective function to be optimized, defaulting value is \code{posterior}.}
}
\value{
A list with two components including \code{posterior} and \code{Y_pred},
where \code{Y_pred} A list with two components, that is,
\describe{
\item{train}{A numeric vector with length(m), each element indicates
the label in the train data point.}
\item{test}{A numeric vector with length(m_new), each element indicates
the label in the test data point.}
}
}
\description{
Fit Gaussian process logistic multinomial regression with GLGP
}
