% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit_gl_regression_gp_rcpp}
\alias{fit_gl_regression_gp_rcpp}
\title{Fit Gaussian process regression with GLGP}
\usage{
fit_gl_regression_gp_rcpp(
  X,
  Y,
  X_new,
  K,
  sigma = 1e-05,
  a2s = NULL,
  threshold = 0.01,
  sparse = TRUE,
  approach = "posterior",
  noise = "same",
  output_cov = FALSE
)
}
\arguments{
\item{X}{Training sample, a (m, d) matrix, each row indicates one point in R^d.}

\item{Y}{A numeric vector with length(m), the training labels.}

\item{X_new}{Testing sample, a (n-m, d) matrix, each row indicates one point in R^d.}

\item{K}{An integer, the number of used eigenpairs to construct heat kernel,
the defaulting value is \code{NULL}, that is, \code{K=min(n,s)}.}

\item{sigma}{A non-negative number, the weight coefficient of ridge penalty on H,
the defaulting value is 1e-3.}

\item{a2s}{A numeric vector, the searching range for bandwidth.}

\item{threshold}{A double, the threshold ratio for sparse GLGP, defaulting value is 0.01.}

\item{sparse}{bool, sparse GLGP or not, defaulting value is \code{TRUE}.}

\item{approach}{A character vector, taking value in c("posterior", "marginal"),
decides which objective function to be optimized, defaulting value is \code{posterior}.}

\item{noise}{A character vector, taking value in c("same", "different"),
indicates whether the noise variances in different locations are different,
defaulting value is \code{same}.}

\item{output_cov}{Bool, whether to output covariance, defaulting value is \code{FALSE}.}
}
\value{
\code{res} A list with three components including \code{pars}, \code{posterior} and \code{Y_pred},
where \code{Y_pred} is also a list with two components, that is,
\describe{
\item{train}{A numeric vector with length(m), each element indicates
the label in the train data point.}
\item{test}{A numeric vector with length(m_new), each element indicates
the label in the test data point.}
}
}
\description{
Fit Gaussian process regression with GLGP
}
