% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit_lae_regression_gp_rcpp}
\alias{fit_lae_regression_gp_rcpp}
\title{Fit Gaussian process regression with local anchor embedding kernels}
\usage{
fit_lae_regression_gp_rcpp(
  X,
  Y,
  X_new,
  s,
  r,
  K = -1,
  sigma = 0.001,
  approach = "posterior",
  models = list(subsample = "kmeans", kernel = "lae", gl = "cluster-normalized", root =
    TRUE),
  output_cov = FALSE,
  nstart = 1
)
}
\arguments{
\item{X}{Training sample, a (m, d) matrix, each row indicates one point in R^d.}

\item{Y}{A numeric vector with length(m), the training labels.}

\item{X_new}{Testing sample, a (n-m, d) matrix, each row indicates one point in R^d.}

\item{s}{An integer indicating the number of the subsampling.}

\item{r}{An integer, the number of the nearest neighbor points.}

\item{K}{An integer, the number of used eigenpairs to construct heat kernel,
the defaulting value is \code{NULL}, that is, \code{K=min(n,s)}.}

\item{sigma}{A non-negative number, the weight coefficient of ridge penalty on H,
the defaulting value is 1e-3.}

\item{approach}{A character vector, taking value in c("posterior", "marginal"),
decides which objective function to be optimized, defaulting value is \code{posterior}.}

\item{models}{A list with four components
\describe{
\item{subsample}{the method of subsampling, the defaulting value is \code{kmeans}.}
\item{kernel}{the type of kernel to compute cross similarity matrix W, the
defaulting value is \code{lae}.}
\item{gl}{the kind of graph Laplacian L, the defaulting value is \code{cluster-normalized}.}
\item{root}{whether to square root eigenvalues of the two steps similarity matrix W,
the defaulting value is \code{TRUE}.}
}}

\item{output_cov}{Bool, whether to output covariance, defaulting value is \code{FALSE}.}

\item{nstart}{Int, the number of random sets chosen in kmeans.}
}
\value{
\code{Y_pred} A numeric vector with length(m_new), each element indicates
the label in the corresponding new sample point.
}
\description{
Fit Gaussian process regression with local anchor embedding kernels
}
\examples{
X <- matrix(runif(6),3,2)
Y <- X[,1]^2 + X[,2]^2
X_new <- matrix(runif(10),5,2)
Y_new <- X_new[,1]^2 + X_new[,2]^2
s <- 6; r <- 3
K <- 5
Y_pred <- fit_lae_regression_gp_rcpp(X, Y, X_new, s, r, K)
}
