% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit_nystrom_logit_gp_rcpp}
\alias{fit_nystrom_logit_gp_rcpp}
\title{Fit Gaussian process logistic regression with Nystrom extension}
\usage{
fit_nystrom_logit_gp_rcpp(
  X,
  Y,
  X_new,
  s,
  K = -1,
  N = NULL,
  sigma = 0.001,
  a2s = NULL,
  approach = "posterior",
  subsample = "kmeans",
  output_cov = FALSE,
  nstart = 1
)
}
\arguments{
\item{X}{Training sample, a (m, d) matrix, each row indicates one point in R^d.}

\item{Y}{A numeric vector with length(m), count of the positive class.}

\item{X_new}{Testing sample, a (n-m, d) matrix, each row indicates one point in R^d.}

\item{s}{An integer indicating the number of the subsampling.}

\item{K}{An integer, the number of used eigenpairs to construct heat kernel,
the defaulting value is \code{NULL}, that is, \code{K=min(n,s)}.}

\item{N}{A numeric vector with length(m), total count.}

\item{sigma}{A non-negative number, the weight coefficient of ridge penalty on H,
the defaulting value is 1e-3.}

\item{a2s}{A numeric vector, the searching range for bandwidth.}

\item{approach}{A character vector, taking value in c("posterior", "marginal"),
decides which objective function to be optimized, defaulting value is \code{posterior}.}

\item{subsample}{The method of subsampling in c(\code{kmeans}, \code{random}, \code{minibatchkmeans}), the defaulting value is \code{kmeans}.}

\item{output_cov}{Bool, whether to output covariance, defaulting value is \code{FALSE}.}

\item{nstart}{Int, the number of random sets chosen in kmeans.}
}
\value{
\code{res} A list with two components including \code{pars}, \code{posterior} and \code{Y_pred},
where \code{Y_pred} is also a list with two components, that is,
\describe{
\item{train}{A numeric vector with length(m), each element indicates
the label in the train data point.}
\item{test}{A numeric vector with length(m_new), each element indicates
the label in the test data point.}
}
}
\description{
Fit Gaussian process logistic regression with Nystrom extension
}
\examples{
X0 <- matrix(rnorm(3*3), 3, 3)
X1 <- matrix(rnorm(3*3, 5), 3, 3)
Y <- c(1,1,1,0,0,0)
X <- rbind(X0,X1)
X0_new <- matrix(rnorm(10*3),10,3)
X1_new <- matrix(rnorm(10*3, 5),10,3)
X_new <- rbind(X0_new, X1_new)
Y_new <- c(rep(1,10),rep(0,10))
s <- 6
K <- 5
Y_pred <- fit_nystrom_logit_gp_rcpp(X, Y, X_new, s, K)
}
