% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit_nystrom_regression_gp_rcpp}
\alias{fit_nystrom_regression_gp_rcpp}
\title{Fit Gaussian process regression with Nystrom extension}
\usage{
fit_nystrom_regression_gp_rcpp(
  X,
  Y,
  X_new,
  s,
  K = -1,
  sigma = 1e-05,
  a2s = NULL,
  approach = "posterior",
  noise = "same",
  subsample = "kmeans",
  output_cov = FALSE,
  nstart = 1
)
}
\arguments{
\item{X}{Training sample, a (m, d) matrix, each row indicates one point in R^d.}

\item{Y}{A numeric vector with length(m), the training labels.}

\item{X_new}{Testing sample, a (n-m, d) matrix, each row indicates one point in R^d.}

\item{s}{An integer indicating the number of the subsampling.}

\item{K}{An integer, the number of used eigenpairs to construct heat kernel,
the defaulting value is \code{NULL}, that is, \code{K=min(n,s)}.}

\item{sigma}{A non-negative number, the weight coefficient of ridge penalty on H,
the defaulting value is 1e-3.}

\item{a2s}{A numeric vector, the searching range for bandwidth.}

\item{approach}{A character vector, taking value in c("posterior", "marginal"),
decides which objective function to be optimized, defaulting value is \code{posterior}.}

\item{noise}{A character vector, taking value in c("same", "different"),
indicates whether the noise variances in different locations are different,
defaulting value is \code{same}.}

\item{subsample}{The method of subsampling in c(\code{kmeans}, \code{random}, \code{minibatchkmeans}), the defaulting value is \code{kmeans}.}

\item{output_cov}{Bool, whether to output covariance, defaulting value is \code{FALSE}.}

\item{nstart}{Int, the number of random sets chosen in kmeans.}
}
\value{
\code{res} A list with three components including \code{pars}, \code{posterior} and \code{Y_pred},
where \code{Y_pred} is also a list with two components, that is,
\describe{
\item{train}{A numeric vector with length(m), each element indicates
the label in the train data point.}
\item{test}{A numeric vector with length(m_new), each element indicates
the label in the test data point.}
}
}
\description{
Fit Gaussian process regression with Nystrom extension
}
\examples{
X <- matrix(runif(6),3,2)
Y <- X[,1]^2 + X[,2]^2
X_new <- matrix(runif(10),5,2)
Y_new <- X_new[,1]^2 + X_new[,2]^2
s <- 6
K <- 5
nystrom <- fit_nystrom_regression_gp_rcpp(X, Y, X_new, s, K)
}
