% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_se_logit_mult_gp.R
\name{fit_se_logit_mult_gp}
\alias{fit_se_logit_mult_gp}
\title{Fit Gaussian process logistic mulotinomial regression with square exponential kernels}
\usage{
fit_se_logit_mult_gp(
  X,
  Y,
  X_new,
  J,
  s,
  r,
  K = NULL,
  N = NULL,
  sigma = 0.001,
  a2s = NULL,
  approach = "posterior",
  cl = NULL,
  models = list(subsample = "kmeans", kernel = "se", gl = "rw", root = FALSE)
)
}
\arguments{
\item{X}{Training sample, a (m, d) matrix, each row indicates one point in R^d.}

\item{Y}{A numeric vector with length(m), count of the positive class.}

\item{X_new}{Testing sample, a (n-m, d) matrix, each row indicates one point in R^d.}

\item{J}{An integer, the number of classes.}

\item{s}{An integer indicating the number of the subsampling.}

\item{r}{An integer, the number of the nearest neighbor points.}

\item{K}{An integer, the number of used eigenpairs to construct heat kernel,
the defaulting value is \code{NULL}, that is, \code{K=s}.}

\item{N}{A numeric vector with length(m), total count.}

\item{sigma}{A non-negative number, the weight coefficient of ridge penalty on H,
the defaulting value is 1e-3.}

\item{a2s}{A numeric vector, the searching range for bandwidth.}

\item{approach}{A character vector, taking value in c("posterior", "marginal"),
decides which objective function to be optimized, defaulting value is \code{posterior}.}

\item{cl}{The cluster to make parallel computing,
typically generated by \code{parallel::makeCluster(num_workers)}.
The defaulting value of cl is NULL, that is, sequential computing.}

\item{models}{A list with four components
\describe{
\item{subsample}{the method of subsampling, the defaulting value is \code{kmeans}.}
\item{kernel}{the type of kernel to compute cross similarity matrix W, the
defaulting value is \code{lae}.}
\item{gl}{the kind of graph Laplacian L, the defaulting value is \code{rw}.}
\item{root}{whether to square root eigenvalues of the two steps similarity matrix W,
the defaulting value is \code{FALSE}.}
}}
}
\value{
\code{Y_pred} A numeric vector with length(m_new), each element indicates
the label in the corresponding new sample point.
}
\description{
Compose J-1 binary logistic regression to implement the multinomial regression.
}
\examples{
X0 <- matrix(rnorm(3*3), 3, 3)
X1 <- matrix(rnorm(3*3, 5), 3, 3)
X2 <- matrix(rnorm(3*3, -5), 3, 3)
Y <- c(rep(0,3), rep(1,3), rep(2,3))
X <- rbind(X0,X1,X2)
X0_new <- matrix(rnorm(10*3),10,3)
X1_new <- matrix(rnorm(10*3, 5),10,3)
X2_new <- matrix(rnorm(10*3, -5),10,3)
X_new <- rbind(X0_new, X1_new, X2_new)
Y_new <- c(rep(0,10),rep(1,10),rep(2,10))
s <- 6; r <- 3
K <- 5
J <- 3
Y_pred <- fit_se_logit_mult_gp(X, Y, X_new, J, s, r, K)
}
