% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphLaplacian.R
\name{graphLaplacian}
\alias{graphLaplacian}
\title{Graph Laplacian type}
\usage{
graphLaplacian(Z, gl, num_class = NULL)
}
\arguments{
\item{Z}{A numeric matrix with dim(n,s), similarity matrix.}

\item{gl}{A character vector, the type of GL.}

\item{num_class}{A numeric vector, the number of points in each cluster}
}
\value{
\code{Z} A numeric matrix with dim(n,s), normalized Z.
}
\description{
Graph Laplacian type
}
\examples{
Z <- matrix(runif(5*2),5,2)
gl <- "rw"
graphLaplacian(Z, gl)
}
