% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{heat_kernel_covariance_rcpp}
\alias{heat_kernel_covariance_rcpp}
\title{Construct a heat kernel covariance matrix by FLGP}
\usage{
heat_kernel_covariance_rcpp(
  X,
  X_new,
  s,
  r,
  t,
  K = -1,
  models = list(subsample = "kmeans", kernel = "lae", gl = "cluster-normalized", root =
    TRUE),
  nstart = 1
)
}
\arguments{
\item{X}{Training sample, a (m, d) matrix, each row indicates one point in R^d.}

\item{X_new}{Testing sample, a (n-m, d) matrix, each row indicates one point in R^d.}

\item{s}{An integer indicating the number of the subsampling.}

\item{r}{An integer, the number of the nearest neighbor points.}

\item{t}{A non-negative number, the heat diffusion time.}

\item{K}{An integer, the number of used eigenpairs to construct heat kernel,
the defaulting value is \code{-1}, that is, \code{K=s}.}

\item{models}{A list, including
\itemize{
\item \code{subsample} the method of subsampling, the defaulting value is \code{kmeans}.
\item \code{kernel} the type of kernel to compute cross similarity matrix W, the
defaulting value is \code{lae}.
\item \code{gl} the kind of graph Laplacian L, the defaulting value is \code{rw}.
\item \code{root} whether to square root eigenvalues of the two steps similarity matrix W,
the defaulting value is \code{FALSE}.
}}

\item{nstart}{An integer, the number of random sets chosen in kmeans,
the defaulting value is \code{1}.}
}
\value{
A numeric matrix with dim (n,m), the heat kernel covariance matrix estimated
by FLGP.
}
\description{
Construct a heat kernel covariance matrix by FLGP
}
\examples{
X <- matrix(rnorm(2*2), 2, 2)
X_new <- matrix(rnorm(3*2),3,2)
s <- 2; r <- 2; t <- 1
heat_kernel_covariance_rcpp(X, X_new, s, r, t)
}
