% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_kernel_covariance.R
\name{heat_kernel_spectrum}
\alias{heat_kernel_spectrum}
\title{Compute spectrum of graph Laplacian by FLAG}
\usage{
heat_kernel_spectrum(
  X,
  X_new,
  s,
  r,
  K = NULL,
  cl = NULL,
  models = list(subsample = "kmeans", kernel = "lae", gl = "rw", root = FALSE)
)
}
\arguments{
\item{X}{Training sample, a (m, d) matrix, each row indicates one point in R^d.}

\item{X_new}{Testing sample, a (n-m, d) matrix, each row indicates one point in R^d.}

\item{s}{An integer indicating the number of the subsampling.}

\item{r}{An integer, the number of the nearest neighbor points.}

\item{K}{An integer, the number of used eigenpairs to construct heat kernel,
the defaulting value is \code{NULL}, that is, \code{K=min(n,s)}.}

\item{cl}{The cluster to make parallel computing,
typically generated by \code{parallel::makeCluster(num_workers)}.
The defaulting value of cl is NULL, that is, sequential computing.}

\item{models}{A list, including
\itemize{
\item \code{subsample} the method of subsampling, the defaulting value is \code{kmeans}.
\item \code{kernel} the type of kernel to compute cross similarity matrix W, the
defaulting value is \code{lae}.
\item \code{gl} the kind of graph Laplacian L, the defaulting value is \code{rw}.
\item \code{root} whether to square root eigenvalues of the two steps similarity matrix W,
the defaulting value is \code{FALSE}.
}}
}
\value{
A list of converged eigenvalues and eigenvectors of W.
}
\description{
Compute spectrum of graph Laplacian by FLAG
}
\examples{
X <- matrix(rnorm(2*2), 2, 2)
X_new <- matrix(rnorm(3*2),3,2)
s <- 2; r <- 2
heat_kernel_spectrum(X, X_new, s, r)
}
