% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_from_spectrum.R
\name{HK_from_spectrum}
\alias{HK_from_spectrum}
\title{Construct heat kernel covariance matrix from the spectrum of W}
\usage{
HK_from_spectrum(eigenpairs, K, t, idx0 = NULL, idx1 = NULL)
}
\arguments{
\item{eigenpairs}{A list of converged eigenvalues and eigenvectors of W,
including \code{values} and \code{vectors}.}

\item{K}{An integer, the number of used eigenpairs.}

\item{t}{A non-negative number, the heat diffusion time.}

\item{idx0, }{An integer vector with length(m0), index vector.}

\item{idx1, }{An integer vector with length(m1), index vector.}
}
\value{
A numeric matrix with dim (m0,m1), the covariance matrix of heat kernel
Gaussian processes.
}
\description{
The graph Laplacian L = I - W.
}
\examples{
Z <- matrix(rnorm(3*3),3,3)
X <- Z\%*\%t(Z)
eigens <- eigen(X)
HK_from_spectrum(eigens, 2, 0.1, c(1:2), c(1:3))
}
