% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSBMGP.R
\name{kappa_vec}
\alias{kappa_vec}
\title{Compute the kappa vector for PG Multinomial inference}
\usage{
kappa_vec(Y, N)
}
\arguments{
\item{Y}{A numeric matrix with dim(m,J), each row is a one hot vector indicating
the label.}

\item{N}{A numeric matrix with dim(m,J-1), each row indicates the count vector.}
}
\value{
\code{kappa} a numeric matrix with dim(m, J-1).
}
\description{
Compute the kappa vector for PG Multinomial inference
}
\examples{
Y <- matrix(sample.int(3*3 ,replace=TRUE),3,3)
kappa_vec(Y, N_vec(Y))
}
