% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KNN_cpp}
\alias{KNN_cpp}
\title{k-nearest neighbor reference points}
\usage{
KNN_cpp(X, U, r = 3L, distance = "Euclidean", output = FALSE, batch = 100L)
}
\arguments{
\item{X}{Original points, a (n,d) matrix, each row indicates one original point.}

\item{U}{Reference points, a (s,d) or (s,d+1) matrix, each row indicates one reference point.}

\item{r}{The number of k-nearest neighbor points, an integer.}

\item{distance}{The distance method to compute k-nearest neighbor points, characters in c("Euclidean", "geodesic"),
including Euclidean distance and geodesic distance, the defaulting distance
is Euclidean distance.}

\item{output}{Bool, whether to output the distance matrix, defaulting value is \code{FALSE}.}

\item{batch}{Int, the batch size, defaulting value is \code{100}.}
}
\value{
If \code{output=FALSE}, \code{list(ind_knn)}, the indexes of KNN, a list with length n, each component of the list is a vector of length r,
indicating the indexes of KNN for the corresponding original point based on the chosen distance.
Otherwise \code{output=TRUE}, \code{list(ind_knn,distances_sp)}, a list with two components, the one is the indexes of KNN,
the other is the sparse distance matrix with dim(n,s).
}
\description{
k-nearest neighbor reference points
}
\examples{
X <- matrix(rnorm(300), nrow=100, ncol=3)
U <- matrix(rnorm(30), nrow=10, ncol=3)
r <- 3
distance <- "Euclidean"
KNN_cpp(X, U, r, distance)
}
