% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lae.R
\name{LAE}
\alias{LAE}
\title{Local anchor embedding}
\usage{
LAE(X, U, r = 3L, gl = "rw", cl = NULL)
}
\arguments{
\item{X}{A numeric matrix with dim (n,d), original sample,
each row indicates one original point in R^d.}

\item{U}{A numeric matrix with dim (s,d) or (s,d+1), sub-sample,
each row indicates one representative point in R^d,
where the d+1 column indicates the number of points in each cluster if it exists.}

\item{r}{An integer, the number of the nearest neighbor points.}

\item{gl}{A character vector in c("rw", "normalized", "cluster-normalized"),
indicates how to construct the stochastic transition matrix. "rw" means random walk,
"normalized" means normalized random walk, "cluster-normalized" means
normalized random walk with cluster membership re-balance. The defaulting gl
is "rw".}

\item{cl}{The cluster to make parallel computing,
typically generated by parallel::makeCluster(num_workers).
The defaulting value of cl is NULL, that is, sequential computing.}
}
\value{
A numeric 'sparse' matrix with dim (n,s) and n*r non-zero entries, r << s,
the stochastic transition matrix from X to U.
}
\description{
Local anchor embedding
}
\examples{
X <- matrix(rnorm(10*3),10,3)
r <- 3
U_1 <- matrix(rnorm(5*3),5,3)
LAE(X, U_1, r)
U_2 <- matrix(rnorm(5*4),5,4)
LAE(X, U_2, r)
}
