% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lae_eigenmap}
\alias{lae_eigenmap}
\title{Compute the Laplacian eigenmap by local anchor embedding}
\usage{
lae_eigenmap(
  X,
  s,
  r = 3L,
  ndim = 2L,
  subsample = "kmeans",
  norm = "cluster-normalized",
  nstart = 1L
)
}
\arguments{
\item{X}{A numeric matrix with dim (n,d),
each row indicates one point in R^d.}

\item{s}{An integer, the number of the induced points.}

\item{r}{An integer, the number of the nearest neighbor points, defaulting \code{3}.}

\item{ndim}{An integer, the dimension of the manifold projection, defaulting \code{2}.}

\item{subsample}{A character vector in c("kmeans", "random"), indicates the method
of subsampling, defaulting "kmeans".}

\item{norm}{A character vector in c("rw", "normalized", "cluster-normalized"),
indicates how to construct the stochastic transition matrix. "rw" means random walk,
"normalized" means normalized random walk, "cluster-normalized" means
normalized random walk with cluster membership re-balance. The defaulting gl
is "cluster-normalized".}

\item{nstart}{An integer, the number of random sets chosen in kmeans, defaulting \code{1}.}
}
\value{
List of two component,
\describe{
\item{eigenvalues}{A numeric vector with length(ndim), the eigenvalues of the graph Laplacian.}
\item{eigenvectors}{A numeric matrix with dim(n,ndim), the eigenvector of the graph Laplacian.
Each row indicates the embedding coordinate representation of one point.}
}
}
\description{
Compute the Laplacian eigenmap by local anchor embedding
}
