% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lae.R
\name{local_anchor_embedding}
\alias{local_anchor_embedding}
\title{Local anchor embedding of one single point by
gradient descent projection with Nesterov's methods}
\usage{
local_anchor_embedding(unit)
}
\arguments{
\item{unit}{A list of x and U:

x is a numeric vector with length d, indicates the single point to be embedded.
U is a numeric matrix with dimension (r, d), the columns of U
are equal to the length of x, including KNN reference points.}
}
\value{
A numeric vector with the length r, convex combination coefficients.
}
\description{
Local anchor embedding of one single point by
gradient descent projection with Nesterov's methods
}
\examples{
x <- rnorm(3)
U <- matrix(rnorm(3*3),3,3)
local_anchor_embedding(list("x"=x,"U"=U))
}
