% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{local_anchor_embedding_cpp}
\alias{local_anchor_embedding_cpp}
\title{Local anchor embedding of one single point by
gradient descent projection with Nesterov's methods}
\usage{
local_anchor_embedding_cpp(x, U)
}
\arguments{
\item{x}{A numeric row vector with length(d), indicates the single point to be embedded.}

\item{U}{A numeric matrix with dimension (r, d), the columns of U
are equal to the length of x, including KNN reference points.}
}
\value{
A numeric vector with the length r, convex combination coefficients.
}
\description{
Local anchor embedding of one single point by
gradient descent projection with Nesterov's methods
}
