% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Compute the log likelihood of the observed}
\usage{
log_likelihood(f, Y)
}
\arguments{
\item{f}{A numeric matrix with dim(M,J-1), latent function.}

\item{Y}{A numeric matrix with dim(m,J), each row is a one hot vector indicating
the label.}
}
\value{
\code{ll} a double, log likelihood value.
}
\description{
Compute the log likelihood of the observed
}
\examples{
f <- matrix(rnorm(3*2),3,2)
Y <- matrix(sample.int(3*3 ,replace=TRUE),3,3)
log_likelihood(f, Y)
}
