% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{marginal_log_likelihood}
\alias{marginal_log_likelihood}
\title{Marginal log likelihood by stick-breaking given C and Y}
\usage{
marginal_log_likelihood(C, Y, N_sample = 100)
}
\arguments{
\item{C}{A numeric matrix with dim(m,m), the self covariance matrix
in the training samples.}

\item{Y}{A numeric matrix with dim(m,J), each row is a one hot vector indicating
the label.}

\item{N_sample}{An integer, the length of the Gibbs sampler chain.}
}
\value{
\code{mll} A double, the marginal log likelihood.
}
\description{
Marginal log likelihood by stick-breaking given C and Y
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- matrix(sample.int(3*3 ,replace=TRUE),3,3)
marginal_log_likelihood(C, Y, 100)
}
