% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_log_likelihood_logit_la.R
\name{marginal_log_likelihood_logit_la}
\alias{marginal_log_likelihood_logit_la}
\title{Mode-finding for binary Laplace GPC with logit link function}
\usage{
marginal_log_likelihood_logit_la(
  C,
  Y,
  N = NULL,
  f = NULL,
  tol = 1e-05,
  max_iter = 100
)
}
\arguments{
\item{C}{A numeric matrix with dim(m,m), covariance matrix.}

\item{Y}{A numeric vector with length(m), count of the positive class.}

\item{N}{A numeric vector with length(m), total count.}

\item{f}{A numeric vector with length(m), initial guess for newton methods,
the defaulting value is \code{NULL}, that is, \code{f=rep(0,m)}.}

\item{tol}{A double, convergence criterion, the defaulting value is \code{1e-5}.}

\item{max_iter}{An integer, the maximum iteration number, defaulting value \code{100}.}
}
\value{
A list with three components.
\describe{
\item{f}{the posterior mode.}
\item{amll}{the Laplace approximation of the marginal log likelihood.}
\item{converge}{0 indicates Newton method succeeds to converge,
1 indicates Newton method fails to converge.}
}
}
\description{
Mode-finding for binary Laplace GPC with logit link function
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- sample(c(0,1), 3,replace=TRUE)
marginal_log_likelihood_logit_la(C, Y)
}
