% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{marginal_log_likelihood_logit_la_cpp}
\alias{marginal_log_likelihood_logit_la_cpp}
\title{Mode-finding for binary Laplace GPC with logit link function}
\usage{
marginal_log_likelihood_logit_la_cpp(C, Y, N, tol = 1e-05, max_iter = 100L)
}
\arguments{
\item{C}{A numeric matrix with dim(m,m), covariance matrix.}

\item{Y}{A numeric vector with length(m), count of the positive class.}

\item{N}{A numeric vector with length(m), total count.}

\item{tol}{A double, convergence criterion, the defaulting value is \code{1e-5}.}

\item{max_iter}{An integer, the maximum iteration number, defaulting value \code{100}.}
}
\value{
\code{amll} A double, the Laplace approximation of the marginal log likelihood.
}
\description{
Mode-finding for binary Laplace GPC with logit link function
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- sample(c(0,1), 3,replace=TRUE)
N <- rep(1,3)
marginal_log_likelihood_logit_la_cpp(C, Y, N)
}
