% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_log_likelihood_reg.R
\name{marginal_log_likelihood_reg}
\alias{marginal_log_likelihood_reg}
\title{Marginal log likelihood in Gaussian regression}
\usage{
marginal_log_likelihood_reg(C_noisy, Y)
}
\arguments{
\item{C_noisy}{A numeric matrix with dim(m,m), noisy covariance matrix,
\verb{C_noisy=C+sigma^2I} .}

\item{Y}{A numeric vector with length(m), response vector.}
}
\value{
\code{mll} A double, the marginal log likelihood value.
}
\description{
Marginal log likelihood in Gaussian regression
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
C_noisy <- C + diag(1e-3, 3)
Y <- rnorm(3)
marginal_log_likelihood_logit_la(C_noisy, Y)
}
