% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_log_likelihood_reg_robust.R
\name{marginal_log_likelihood_reg_robust}
\alias{marginal_log_likelihood_reg_robust}
\title{Robust marginal log likelihood in Gaussian regression}
\usage{
marginal_log_likelihood_reg_robust(eigenpair, Y, theta, K)
}
\arguments{
\item{eigenpair}{A list includes values and vectors.}

\item{Y}{A numeric vector with length(m), response vector.}

\item{theta}{A numeric vector, hyper-parameters.}

\item{K}{An integer, the number of used eigenpairs.}
}
\value{
\code{mll} A double, the marginal log likelihood value.
}
\description{
Robust marginal log likelihood in Gaussian regression
}
\examples{
X <- matrix(rnorm(3*3), 3, 3)
Y <- rowSums(X)
X_new <- matrix(rnorm(10*3),10,3)
s <- 5; r <- 3
m <- 3; K <- 3
eigenpair <- heat_kernel_spectrum(X, X_new, s, r, K=K)
theta <- c(1,0.1)
K <- 3
marginal_log_likelihood_reg_robust(eigenpair, Y, theta, K)
}
