% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{marginal_log_likelihood_regression_cpp}
\alias{marginal_log_likelihood_regression_cpp}
\title{Marginal log likelihood function for GPR}
\usage{
marginal_log_likelihood_regression_cpp(C, Y)
}
\arguments{
\item{C}{A numeric matrix with dim(m,m), covariance matrix.}

\item{Y}{A numeric vector with length(m), count of the positive class.}
}
\value{
\code{mll} A double, the marginal log likelihood.
}
\description{
Marginal log likelihood function for GPR
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- runif(3)
marginal_log_likelihood_regression_cpp(C, Y)
}
