% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multi_train_split_cpp}
\alias{multi_train_split_cpp}
\title{split one multi-classes data set into n_classes binary-classes data set
The multi-class labels should be continuous integers starting from \code{0}, such as "0,1,2,...,9".}
\usage{
multi_train_split_cpp(Y)
}
\arguments{
\item{Y}{A numeric vector with length(m), indicating the labels of multi-classes.}
}
\value{
A numeric matrix with dim(m,n_classes).
}
\description{
split one multi-classes data set into n_classes binary-classes data set
The multi-class labels should be continuous integers starting from \code{0}, such as "0,1,2,...,9".
}
