% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSBMGP.R
\name{N_vec}
\alias{N_vec}
\title{Compute the count vector for PG Multinomial inference}
\usage{
N_vec(Y)
}
\arguments{
\item{Y}{A numeric matrix with dim(m,J), each row is a one hot vector indicating
the label.}
}
\value{
\code{N} a numeric matrix with dim(m,J-1), each row indicates the count vector.
}
\description{
Compute the count vector for PG Multinomial inference
}
\examples{
Y <- matrix(sample.int(3*3 ,replace=TRUE),3,3)
N_vec(Y)
}
