% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negative_log_likelihood}
\alias{negative_log_likelihood}
\title{compute the negative log likelihood of test data}
\usage{
negative_log_likelihood(mean, cov, target, type)
}
\arguments{
\item{mean}{Posterior mean.}

\item{cov}{Posterior covariance.}

\item{target}{Targeted labels.}

\item{type}{The task type, \code{regression}, \code{binary} or \code{multinomial}.}
}
\value{
negative log likelihood.
}
\description{
compute the negative log likelihood of test data
}
