% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_log_posterior_logit.R
\name{negative_log_posterior_logit}
\alias{negative_log_posterior_logit}
\title{Create negative log posterior functional with inverse gamma prior for logistic regression}
\usage{
negative_log_posterior_logit(
  eigenpair,
  Y,
  idx,
  K,
  N = NULL,
  sigma = 0.001,
  p = 0.01,
  q = 10,
  tau = 2
)
}
\arguments{
\item{eigenpair}{A list includes values and vectors.}

\item{Y}{A numeric vector with length(m), count of the positive class.}

\item{idx}{An integer vector with length(m), the index of training samples.}

\item{K}{An integer, the number of used eigenpairs.}

\item{N}{A numeric vector with length(m), total count.}

\item{sigma}{A non-negative number, the weight coefficient of ridge penalty on H,
the defaulting value is 1e-3.}

\item{p}{A positive double, hyper-parameter, defaulting value is 1e-2.}

\item{q}{A positive double, hyper-parameter, defaulting value is 10.}

\item{tau}{A positive double, hyper-parameter, defaulting value is 2.}
}
\value{
\code{nlp} negative log posterior functional with parameter t.
}
\description{
Create negative log posterior functional with inverse gamma prior for logistic regression
}
\examples{
Z <- matrix(rnorm(3*3),3,3)
X <- Z\%*\%t(Z)
eigenpair <- eigen(X)
Y <- sample(c(0,1),3, replace=TRUE)
negative_log_posterior_logit(eigenpair, Y, c(1:3), 2)
}
