% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_log_posterior_reg.R
\name{negative_log_posterior_reg}
\alias{negative_log_posterior_reg}
\title{Create negative log posterior functional with inverse gamma prior for regression}
\usage{
negative_log_posterior_reg(eigenpair, Y, K, p = 0.01, q = 10, tau = 2)
}
\arguments{
\item{eigenpair}{A list includes values and vectors.}

\item{Y}{A numeric vector with length(m), response vector.}

\item{K}{An integer, the number of used eigenpairs.}

\item{p}{A positive double, hyper-parameter, defaulting value is 1e-2.}

\item{q}{A positive double, hyper-parameter, defaulting value is 10.}

\item{tau}{A positive double, hyper-parameter, defaulting value is 2.}
}
\value{
\code{nlp} negative log posterior functional with parameter theta,
where theta = c(t, sigma).
}
\description{
Create negative log posterior functional with inverse gamma prior for regression
}
\examples{
Z <- matrix(rnorm(3*3),3,3)
X <- Z\%*\%t(Z)
eigenpair <- eigen(X)
Y <- rnorm(3)
K <- 2
negative_log_posterior_reg(eigenpair, Y, K)
}
