% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_marginal_likelihood_logit.R
\name{negative_marginal_likelihood_logit}
\alias{negative_marginal_likelihood_logit}
\title{Create negative log marginal likelihood functional}
\usage{
negative_marginal_likelihood_logit(
  eigenpair,
  Y,
  idx,
  K,
  N = NULL,
  sigma = 0.001
)
}
\arguments{
\item{eigenpair}{A list includes values and vectors.}

\item{Y}{A numeric vector with length(m), count of the positive class.}

\item{idx}{An integer vector with length(m), the index of training samples.}

\item{K}{An integer, the number of used eigenpairs.}

\item{N}{A numeric vector with length(m), total count.}

\item{sigma}{A non-negative number, the weight coefficient of ridge penalty on H,
the defaulting value is 1e-3.}
}
\value{
\code{nll} negative log marginal likelihood functional with parameter t.
}
\description{
Create negative log marginal likelihood functional
}
\examples{
Z <- matrix(rnorm(3*3),3,3)
X <- Z\%*\%t(Z)
eigenpair <- eigen(X)
Y <- sample(c(0,1),3, replace=TRUE)
negative_marginal_likelihood_logit(eigenpair, Y, c(1:3), 2)
}
