% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_marginal_likelihood_reg.R
\name{negative_marginal_likelihood_reg}
\alias{negative_marginal_likelihood_reg}
\title{Create negative log marginal likelihood functional for regression}
\usage{
negative_marginal_likelihood_reg(eigenpair, Y, K)
}
\arguments{
\item{eigenpair}{A list includes values and vectors.}

\item{Y}{A numeric vector with length(m), response vector.}

\item{K}{An integer, the number of used eigenpairs.}
}
\value{
\code{nll} negative log marginal likelihood functional with parameter theta,
where theta = c(t, sigma).
}
\description{
Create negative log marginal likelihood functional for regression
}
\examples{
Z <- matrix(rnorm(3*3),3,3)
X <- Z\%*\%t(Z)
eigenpair <- eigen(X)
Y <- rnorm(3)
K <- 2
negative_marginal_likelihood_reg(eigenpair, Y, K)
}
