% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGGibbsSampler.R
\name{PG_Gibbs_sampler}
\alias{PG_Gibbs_sampler}
\title{Gibbs sampler for Polya-gamma auxiliary variables}
\usage{
PG_Gibbs_sampler(aug_data = AugmentedData(), N_sample = 100)
}
\arguments{
\item{aug_data}{An augmented data for polya-gamma sampling.}

\item{N_sample}{An integer, the length of the sampling chain.}
}
\value{
\code{aug_data} The updated augmented data.
}
\description{
Gibbs sampler for Polya-gamma auxiliary variables
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- matrix(sample.int(3*3 ,replace=TRUE),3,3)
aug_data = AugmentedData(C,Y)
PG_Gibbs_sampler(aug_data, 100)
}
