% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_breaking.R
\name{pi_to_f}
\alias{pi_to_f}
\title{Stick breaking transform from pi to f}
\usage{
pi_to_f(pi)
}
\arguments{
\item{pi}{A numeric matrix with dim(M,J), class probability.}
}
\value{
\code{f} a numeric matrix with dim(M,J-1), latent function.
}
\description{
Stick breaking transform from pi to f
}
\examples{
pi <- matrix(abs(rnorm(3*3))+1,3,3)
pi <- Matrix::rowScale(pi, Matrix::rowSums(pi)^{-1})
pi_to_f(pi)
}
