% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posterior_distribution_classification}
\alias{posterior_distribution_classification}
\title{compute the posterior distribution of test data in GPC}
\usage{
posterior_distribution_classification(
  C11,
  C21,
  C22,
  Y,
  tol = 1e-05,
  max_iter = 100L
)
}
\arguments{
\item{C11}{Prior covariance matrix on training data.}

\item{C21}{Prior covariance matrix between test data and training data.}

\item{C22}{Prior covariance matrix on test data.}

\item{Y}{Training labels.}
}
\description{
compute the posterior distribution of test data in GPC
}
