% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_f}
\alias{resample_f}
\title{Resample f given omega and y}
\usage{
resample_f(aug_data = AugmentedData())
}
\arguments{
\item{aug_data}{An augmented data for polya-gamma sampling.}
}
\value{
\code{f} A numeric matrix with dim(m,J-1), the updated f.
}
\description{
Resample f given omega and y
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- matrix(sample.int(3*3 ,replace=TRUE),3,3)
aug_data = AugmentedData(C,Y)
resample_f(aug_data)
}
