% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_omega}
\alias{resample_omega}
\title{Resample omega given N and f}
\usage{
resample_omega(aug_data = AugmentedData())
}
\arguments{
\item{aug_data}{An augmented data for polya-gamma sampling.}
}
\value{
\code{omega} A numeric matrix with dim(m, J-1), the updated omega.
}
\description{
Resample omega given N and f
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- matrix(sample.int(3*3 ,replace=TRUE),3,3)
aug_data = AugmentedData(C,Y)
resample_omega(aug_data)
}
