% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMultinomialGP.R
\name{SBMultinomialGP}
\alias{SBMultinomialGP}
\title{Construct stick breaking multinomial Gaussian process models}
\usage{
SBMultinomialGP(C, Y)
}
\arguments{
\item{C}{A numeric matrix with dim(m, m), the covariance matrix of training samples,
where \code{C[i,j]} indicates the covariance between \code{X_i} and \code{X_j}.}

\item{Y}{A numeric matrix with dim(m,J), each row is a one hot vector indicating
the label.}
}
\value{
An instance of the class \code{SBMultinomialGP}.
}
\description{
Construct stick breaking multinomial Gaussian process models
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- diag(rep(1,3))
SBMultinomialGP(C,Y)
}
