% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_from_Z.R
\name{spectrum_from_Z}
\alias{spectrum_from_Z}
\title{The spectrum of the similarity matrix W from Z}
\usage{
spectrum_from_Z(Z, K = NULL, root = FALSE)
}
\arguments{
\item{Z}{A numeric matrix with dim (n,s), the cross similarity matrix
between the original sample and the sub-sample.}

\item{K}{An integer, the number of eigenpairs requested, the defaulting value
is NULL, indicating all non-trivial eigenpairs, that is, K=min(n,s).}

\item{root}{A logical value, indicating whether to square root eigenvalues of W,
the defaulting value is FALSE.}
}
\value{
A list of converged eigenvalues and eigenvectors of W.
}
\description{
The transition matrix W is the two step composition of
the cross similarity matrix Z by $W=ZLambda^{-1}Z^T$.
}
\examples{
Z <- matrix(abs(rnorm(5*3)),5,3)
Z <- Matrix::rowScale(Z, Matrix::rowSums(Z)^{-1})
spectrum_from_Z(Z)
}
