% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spectrum_from_Z_cpp}
\alias{spectrum_from_Z_cpp}
\title{The spectrum of the similarity matrix W from Z}
\usage{
spectrum_from_Z_cpp(Z, K = -1L, root = FALSE)
}
\arguments{
\item{Z}{A numeric matrix with dim (n,s), the cross similarity matrix
between the original sample and the sub-sample.}

\item{K}{An integer, the number of eigenpairs requested, the defaulting value
is NULL, indicating all non-trivial eigenpairs, that is, K=min(n,s).}

\item{root}{A logical value, indicating whether to square root eigenvalues of W,
the defaulting value is FALSE.}
}
\value{
A list of converged eigenvalues and eigenvectors of W.
\describe{
\item{values}{eigenvalues, descending order.}
\item{vectors}{eigenvectors, the vectors are normalized to sqrt(n) length.}
}
}
\description{
The transition matrix W is the two step composition of
the cross similarity matrix Z by $W=ZLambda^{-1}Z^T$.
}
\examples{
Z <- Matrix::sparseMatrix(i=c(1:5),j=sample.int(5),x=abs(rnorm(5)),repr = "R")
# Z <- diag(1/Matrix::rowSums(Z))\%*\%Z
K <- 2
spectrum_from_Z_cpp(Z, K)
}
