% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pgbinary.R
\name{test_pgbinary}
\alias{test_pgbinary}
\title{Predict labels on new samples with Polya-Gamma}
\usage{
test_pgbinary(C, Y, Cnv, N = NULL, N_sample = 1000, full_out = FALSE)
}
\arguments{
\item{C}{A numeric matrix with dim(m,m), the self covariance matrix
in the training samples.}

\item{Y}{A numeric vector with length(m), count of the positive class.}

\item{Cnv}{A numeric matrix with dim(m_new,m), cross covariance matrix
between new sample and training sample.}

\item{N}{A numeric vector with length(m), total count.}

\item{N_sample}{An integer, the length of the Gibbs sampler chain.}

\item{full_out}{Bool, whether or not to output pi_new, defaulting value is \code{FALSE}.}
}
\value{
\code{Y_new} if \code{full_out=FALSE}, otherwise \code{list(pi_new,Y_new)}.
}
\description{
Predict labels on new samples with Polya-Gamma
}
\examples{
Z <- matrix(rnorm(3*3),3,3)
C <- Z\%*\%t(Z)
Y <- sample(c(0,1),3, replace=TRUE)
Cnv <- matrix(rnorm(5*3),5,3)
test_pgbinary(C, Y, Cnv)
}
