% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_pgbinary_cpp}
\alias{test_pgbinary_cpp}
\title{Predict labels on new samples with Polya-Gamma}
\usage{
test_pgbinary_cpp(C, Y, Cnv, N_sample = 100L, output_pi = FALSE)
}
\arguments{
\item{C}{A numeric matrix with dim(m,m), the self covariance matrix
in the training samples.}

\item{Y}{A numeric vector with length(m), count of the positive class.}

\item{Cnv}{A numeric matrix with dim(m_new,m), cross covariance matrix
between new sample and training sample.}

\item{N_sample}{An integer, the length of the Gibbs sampler chain.}

\item{output_pi}{Bool, whether or not to output pi_new, defaulting value is \code{FALSE}.}
}
\value{
\code{list(Y_pred)} if \code{output_pi=FALSE}, otherwise \code{list(Y_pred,pi_pred)}.
}
\description{
Predict labels on new samples with Polya-Gamma
}
\examples{
Z <- matrix(rnorm(3*3),3,3)
C <- Z\%*\%t(Z)
Y <- sample(c(0,1),3, replace=TRUE)
Cnv <- matrix(rnorm(5*3),5,3)
test_pgbinary_cpp(C, Y, Cnv)
}
