% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pgmult.R
\name{test_pgmult_label}
\alias{test_pgmult_label}
\title{Predict labels on new samples}
\usage{
test_pgmult_label(C, Y, Cnv, N_sample = 1000)
}
\arguments{
\item{C}{A numeric matrix with dim(m,m), the self covariance matrix
in the training samples}

\item{Y}{A numeric matrix with dim(m,J), each row is a one hot vector indicating
the label.}

\item{Cnv}{A numeric matrix with dim(m_new,m), cross covariance matrix
between new sample and training sample.}

\item{N_sample}{An integer, the length of the Gibbs sampler chain.}
}
\value{
\code{Y_new} A numeric matrix with dim(m_new,J), each row indicates
the label in the corresponding new sample point.
}
\description{
Predict labels on new samples
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- matrix(sample.int(3*3 ,replace=TRUE),3,3)
Cnv <- matrix(rnorm(5*3),5,3)
test_pgmult_label(C, Y, Cnv, 1000)
}
