% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_reg.R
\name{test_reg}
\alias{test_reg}
\title{Predict response on new samples for Gaussian regression}
\usage{
test_reg(C, Y, Cnv)
}
\arguments{
\item{C}{A numeric matrix with dim(m,m), the noisy self covariance matrix
in the training samples.}

\item{Y}{A numeric vector with length(m), response vector.}

\item{Cnv}{A numeric matrix with dim(m_new,m), cross covariance matrix
between new sample and training sample.}
}
\value{
\code{Y_pred} A numeric vector with length(m_new), predictive response vector.
}
\description{
Predict response on new samples for Gaussian regression
}
\examples{
Z <- matrix(rnorm(3*3),3,3)
C <- Z\%*\%t(Z)
Y <- rnorm(3)
Cnv <- matrix(rnorm(5*3),5,3)
test_pgbinary(C, Y, Cnv)
}
