% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_reg_robust.R
\name{test_reg_robust}
\alias{test_reg_robust}
\title{Robustly predict response on new samples for Gaussian regression}
\usage{
test_reg_robust(eigenpair, Y, theta, K)
}
\arguments{
\item{eigenpair}{A list includes values and vectors.}

\item{Y}{A numeric vector with length(m), response vector.}

\item{theta}{A numeric vector, hyper-parameters.}

\item{K}{An integer, the number of used eigenpairs.}
}
\value{
\code{Y_pred} A numeric vector with length(m_new), predictive response vector.
}
\description{
Robustly predict response on new samples for Gaussian regression
}
\examples{
X <- matrix(rnorm(3*3), 3, 3)
Y <- rowSums(X)
X_new <- matrix(rnorm(10*3),10,3)
s <- 5; r <- 3
m <- 3; K <- 3
eigenpair <- heat_kernel_spectrum(X, X_new, s, r, K=K)
theta <- c(1,0.1)
K <- 3
test_reg_robust(eigenpair, Y, theta, K)
}
