% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_regression_cpp}
\alias{test_regression_cpp}
\title{Predict labels on new samples in the regression}
\usage{
test_regression_cpp(C, Y, Cnv)
}
\arguments{
\item{C}{A numeric matrix with dim(m,m), the self covariance matrix of noisy targets Y.
in the training samples.}

\item{Y}{A numeric vector with length(m), the training labels.}

\item{Cnv}{A numeric matrix with dim(m_new,m), cross covariance matrix
between new sample and training sample.}
}
\value{
\code{Y_pred}, A numeric vector with length(m_new), the prediction labels.
}
\description{
Predict labels on new samples in the regression
}
\examples{
Z <- matrix(rnorm(3*3),3,3)
C <- Z\%*\%t(Z)
Y <- runif(3)
Cnv <- matrix(rnorm(5*3),5,3)
test_regression_cpp(C, Y, Cnv)
}
