% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_lae_logit_gp.R
\name{train_lae_logit_gp}
\alias{train_lae_logit_gp}
\title{Learn diffusion time t by maximizing log marginal likelihood}
\usage{
train_lae_logit_gp(
  eigenpair,
  Y,
  m,
  K,
  sigma = 0.001,
  N = NULL,
  t0 = NULL,
  lower = 0,
  upper = 100
)
}
\arguments{
\item{eigenpair}{A list includes values and vectors.}

\item{Y}{A numeric vector with length(m), count of the positive class.}

\item{m}{An integer, the number of labeled samples.}

\item{K}{An integer, the number of used eigenpairs.}

\item{sigma}{A non-negative number, the weight coefficient of ridge penalty on H,
the defaulting value is 1e-3.}

\item{N}{A numeric vector with length(m), total count.}

\item{t0}{A positive double, the initial guess for t, defaulting value \code{10}.}

\item{lower}{the lower bound on t, defaulting value \code{0}.}

\item{upper}{the upper bound on t, defaulting value \code{100}.}
}
\value{
A list with two components
\describe{
\item{t}{the optimal diffusion time.}
\item{mll}{the corresponding optimal log marginal likelihood value.}
}
}
\description{
Learn diffusion time t by maximizing log marginal likelihood
}
\examples{
X0 <- matrix(rnorm(3*3), 3, 3)
X1 <- matrix(rnorm(3*3, 5), 3, 3)
Y <- c(1,1,1,0,0,0)
X <- rbind(X0,X1)
X0_new <- matrix(rnorm(10*3),10,3)
X1_new <- matrix(rnorm(10*3, 2),10,3)
X_new <- rbind(X0_new, X1_new)
s <- 6; r <- 3
m <- 6; K <- 3
eigenpair <- heat_kernel_spectrum(X, X_new, s, r, K=K)
train_lae_logit_gp(eigenpair, Y, m, K)
}
