% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_svd.R
\name{truncated_SVD}
\alias{truncated_SVD}
\title{Truncated SVD, compute the non-trivial spectrums of A\%\\emph{\%t(A) by calculating the spectrums
of t(A)\%}\%A firstly.}
\usage{
truncated_SVD(A, K = NULL)
}
\arguments{
\item{A}{A (n,s) numeric matrix, sparseMatrix is supported as well.
The eigenvalues and eigenvectors of A\%*\%t(A) will be computed.}

\item{K}{An integer, the number of eigenvalues requested.}
}
\value{
A list of converged eigenvalues and eigenvectors of AA^T. The return results
should be the same as svd(A, nu=K, nv=0). When ncol(A) << nrow(A), truncated_svd
will be much faster than svd:
\describe{
\item{values}{a vector containing the K eigenvalues of AA^T, sorted in decreasing order.}
\item{vectors}{a n \* K matrix whose columns contain the eigenvectors of AA^T.
The vectors are normalized to unit length.}
}
}
\description{
Truncated SVD, compute the non-trivial spectrums of A\%\\emph{\%t(A) by calculating the spectrums
of t(A)\%}\%A firstly.
}
\examples{
A <- matrix(rnorm(5*2), 5, 2)
truncated_SVD(A, 2)
}
