% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{v_to_z_cpp}
\alias{v_to_z_cpp}
\title{Simplex projection into convex combination coefficients}
\usage{
v_to_z_cpp(v)
}
\arguments{
\item{v}{A numeric vector to be projected.}
}
\value{
A numeric vector of convex combination coefficients with
the same length of v.
}
\description{
Simplex projection into convex combination coefficients
}
\examples{
v <- rnorm(3)
v_to_z_cpp(v)
}
