% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{which_minn}
\alias{which_minn}
\title{Find indexes of the first r smallest elements in arrays based on Bubblesorting}
\usage{
which_minn(z, r)
}
\arguments{
\item{z}{a vector of length s, including elements to be sorted.}

\item{r}{The number of the smallest elements, an integer.}
}
\value{
A vector of length r, indicating the indexes of the first r smallest elements.
}
\description{
Find indexes of the first r smallest elements in arrays based on Bubblesorting
}
\examples{
z <- c(1,3,2)
r <- 2
which_minn(z, r)
}
