% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{which_minn_rcpp}
\alias{which_minn_rcpp}
\title{Find indexes of the first r smallest elements in arrays based on Bubblesorting}
\usage{
which_minn_rcpp(z, r)
}
\arguments{
\item{z}{Numeric vector, including elements to be sorted.}

\item{r}{Int, the number of the smallest elements.}
}
\value{
Integer vector of length r, indicating the indexes of the first r smallest elements.
}
\description{
Find indexes of the first r smallest elements in arrays based on Bubblesorting
}
\examples{
z <- c(1,3,2)
r <- 2
which_minn_rcpp(z, r)
}
