% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentedData.R
\name{AugmentedData}
\alias{AugmentedData}
\title{Construct stick breaking multinomial Gaussian process models}
\usage{
AugmentedData(C, Y, inverse = FALSE)
}
\arguments{
\item{C}{A numeric matrix with dim(m, m), the covariance matrix of training samples,
where \code{C[i,j]} indicates the covariance between \code{X_i} and \code{X_j}.}

\item{Y}{A numeric matrix with dim(m,J), each row is a one hot vector indicating
the label.}

\item{inverse}{Bool, whether to compute C inverse firstly, defaulting value is FALSE.}
}
\value{
An instance of the class \code{AugmentedData}.
}
\description{
Construct stick breaking multinomial Gaussian process models
}
\examples{
A <- matrix(rnorm(3*3),3,3)
C <- A\%*\%t(A)
Y <- matrix(sample.int(3*3 ,replace=TRUE),3,3)
AugmentedData(C,Y)
}
