% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_from_spectrum.R
\name{HK_from_spectrum}
\alias{HK_from_spectrum}
\title{Construct heat kernel covariance matrix from the spectrum of W}
\usage{
HK_from_spectrum(eigenpairs, m, K, t)
}
\arguments{
\item{eigenpairs}{A list of converged eigenvalues and eigenvectors of W,
including \code{values} and \code{vectors}.}

\item{m}{An integer, the number of labeled samples.}

\item{K}{An integer, the number of used eigenpairs.}

\item{t}{A non-negative number, the heat diffusion time.}
}
\value{
A numeric matrix with dim (n,m), the covariance matrix of heat kernel
Gaussian processes.
}
\description{
The graph Laplacian L = I - W.
}
\examples{
Z <- matrix(rnorm(3*3),3,3)
X <- Z\%*\%t(Z)
eigens <- eigen(X)
HK_from_spectrum(eigens, 2, 2, 0.1)
}
